feat: Add file operations component for loading and saving card designs
- Implemented FileOperations component to handle design file operations.
- Added functionality to load and save card designs in JSON format.
- Integrated file operations into the main page for user accessibility.

chore: Update package.json for deployment

- Added gh-pages dependency for deployment.
- Included deploy script in package.json.

fix: Adjust card duplication component styling

- Removed unnecessary margin-top from CardDuplication component.

feat: Introduce text field style editor

- Created TextFieldStyleEditor component for editing text styles.
- Integrated text style editing into the main design interface.

refactor: Enhance serial number control with text styles

- Updated SerialNumberControl to accept text styles and handle style changes.
- Integrated TextStyleFieldGroup for better organization of text style controls.

chore: Add CNAME for custom domain

- Created CNAME file for custom domain configuration.

fix: Update range input handling to support decimal values

- Changed input parsing from parseInt to parseFloat for better value handling.

style: Improve global styles for print layout

- Adjusted styles for print media to hide unnecessary elements.
- Enhanced layout for better print presentation.

chore: Remove unused API route

- Deleted hello.ts API route as it was not utilized in the application.

feat: Add Google site verification meta tag

- Included Google site verification meta tag in the document head for SEO purposes.

+++ src/components/Range.tsx
@@ -20,7 +20,7 @@ function BaseRange({ label, value, min, max, onChange }: BaseRangeProps) {
           max={max}
           value={value}
           onChange={(e) => {
-            const newValue = parseInt(e.target.value);
+            const newValue = parseFloat(e.target.value);
             if (!isNaN(newValue)) {
               onChange(newValue);
             }
@@ -31,7 +31,7 @@ function BaseRange({ label, value, min, max, onChange }: BaseRangeProps) {
           type="number"
           value={value}
           onChange={(e) => {
-            const newValue = parseInt(e.target.value);
+            const newValue = parseFloat(e.target.value);
             if (!isNaN(newValue)) {
               onChange(newValue);
             }
